don't increment ref_done/nm_mem.refcount on error
We do not want to increment ref_done if the memory allocation
finalizazion failed, otherwise a second call to netmap_get_memory
would falsely assume that nothing needs to be done (in the current
implementation it also assumes the allocator is good, which later
leads to kernel panic).

For consistency, we also do not increment the refcount in case of
error.