[linux] keep a ref to drivers' modules
Keeping a reference to the net_device is apparently not sufficent for
us. If the driver module is unloaded, some cleanup is always performed
before waiting for all net_devices to become free.  In particular, rings are
deallocated. This causes a kernel ops in *_netmap_rxsync or _txsync
if we were polling that device.

This patch is not necessarily the best way to solve the issue, but
it works.