new schema for linux drivers patches
The linux patches are now autogenerated (using the scripts in the scripts/
directory) from the netmap-* branches in our linux.git.

The old 'patches' directory is now a symbolic link to either 'tmp-patches' or
'final-patches'.

- 'tmp-patches' is obtained with
   "scritps/np get-range <first included vers> <first excluded vers>"
   Each patch in 'tmp-patches' applies to a single driver for a single
   version of linux.

- 'final-patches' is obtained with "scripts/np minimize-all". These
   patches are the result of a repeated join process on the patches in
   'tmp-patches', in order to find a minimal set of patches that still
   produces the same files for each linux version in the range.

   If V is the next, not yet available linux version, the patches in
   'final-patches' that are valid up to version V (excluded) may be
   optimistically extended to infinity using "scripts/np infty-all V".

NOTE: the old mellanox patches are retained in the wip-patches directory
for now, since they don't apply cleanly to any version of linux that I
have tried. They have not been included in any netmap-* branch yet.