new NIOCGINFO with no (visible) side effects
NIOCGINFO now returns the memory size according to the current
configuration of the relevant memory allocator, which is a global
allocator for real interfaces, or a private one for VALE ports. A
NIOCGINFO issued with no interface name reports about the global
allocator.

Please note that this information is accurate only if the relevant
allocator is already in use by some other means (NIOCREGIF or mmap),
otherwise it may be subject to change.