call nm_register even if ifp is null
virtual adapters are going to have a null na->ifp, but
we still need to call their nm_register callback.
On the other end, we cannot call the nm_register callback
of hw adapters that have na->ifp == NULL, since that means that
the nic driver module has been removed, and the callback
is simply not there anymore.